loggen: First iteration of the program
This is the initial shot for the loggen program. It can connect to
INET or Unix domain sockets both in stream or datagram mode.

The tool can create multiple active threads chich are sending the
actual syslog messages but also idle threads just connecting to the
remote host without sending any data.

Messages to be sent can be passed on the command line, on the standard
input or by reading a file.

Signed-off-by: Jimmy Thrasibule <jimmy@thrasibule.mx>